+++ src/auto-evo/simulation/selection_pressure/SelectionPressure.cs
@@ -9,70 +9,6 @@ public abstract class SelectionPressure
     public List<IMutationStrategy<EarlyMulticellularSpecies>> MulticellularMutations;
     public int EnergyProvided = 0;
 
-    public static List<SelectionPressure> PressuresFromPatch(Species? species, Patch patch, MutationLibrary partList, SimulationCache cache, List<SelectionPressure> niche)
-    {
-        // Add Selection pressures
-        // TODO: move this somewhere else
-        var selectionPressures = new List<SelectionPressure>();
-
-        selectionPressures.AddRange(niche);
-
-        selectionPressures.Add(new OsmoregulationEfficiencyPressure(patch, 0.5f));
-
-        return selectionPressures;
-    }
-
-    public static List<Miche> AutotrophicMichesForPatch(Patch patch, SimulationCache cache)
-    {
-        var retval = new List<Miche>();
-
-        // Hydrogen Sulfide
-        if (patch.GetCompoundAmount("hydrogensulfide") > 0)
-        {
-            retval.Add(
-                new Miche("Hydrogen Sulfide Chemosynthesis", new AutotrophEnergyEfficiencyPressure(patch, Compound.ByName("hydrogensulfide"), 10.0f),
-                    new List<Miche> { new Miche("Mobile Hydrogen Sulfide Chemosynthesis", new ReachCompoundCloudPressure(1.0f), 
-                        new List<Miche> { new Miche("and don't die", new MetabolicStabilityPressure(patch, 1.0f)) }) 
-                    })
-            );
-        }
-
-        // Sunlight
-        if (patch.GetCompoundAmount("sunlight") > 0)
-        {
-            retval.Add(
-                new Miche("Photosynthesis", new AutotrophEnergyEfficiencyPressure(patch, Compound.ByName("sunlight"), 1.0f),
-                    new List<Miche> { 
-                        new Miche("and store energy for night", new StoreGlucosePressure(50.8f), new List<Miche>{ new Miche("and don't die", new MetabolicStabilityPressure(patch, 1.0f)) }),
-                        new Miche("and don't die", new MetabolicStabilityPressure(patch, 1.0f))
-                    })
-            );
-        }
-
-        // Glucose
-        if (patch.GetCompoundAmount("glucose") > 0)
-        {
-            retval.Add(new Miche("Glucose Consumption", new ReachCompoundCloudPressure(5.0f),
-                new List<Miche> { new Miche("and don't die", new MetabolicStabilityPressure(patch, 1.0f)) }
-            ));
-        }
-
-        return retval;
-    }
-    
-    public static List<Miche> PredationMiches(Patch patch, HashSet<Species> prey, SimulationCache cache)
-    {
-        var retval = new List<Miche>();
-
-        foreach (var possiblePrey in prey)
-        {
-            var pressure = new PredationEffectivenessPressure(possiblePrey, 10.0f);
-            retval.Add(new Miche(pressure.Name(), pressure, new List<Miche> { new Miche("and don't die", new MetabolicStabilityPressure(patch, 1.0f)) }));
-        }
-
-        return retval;
-    }
-
     public SelectionPressure(bool exclusive, float strength, List<IMutationStrategy<MicrobeSpecies>> microbeMutations, List<IMutationStrategy<EarlyMulticellularSpecies>> multicellularMutations)
     {
         Strength = strength;
@@ -97,7 +33,7 @@ public abstract class SelectionPressure
 
         if (score1 == 0)
         {
-            return -999.0f;
+            return Miche.INVIABLE_PRESSURE_RESULT;
         }
 
         if (score2 == 0)
